fix: decode artifact_scores NUMERIC as f64 via ::float8 cast
Les colonnes freshness/adoption/reliability/abandonment/overall de
artifact_scores sont NUMERIC(4,3). sqlx refuse de les décoder en
Option<f64> (mismatch FLOAT8 vs NUMERIC), et l'API renvoyait une
500 dès qu'un score existait.

Bug latent dans services/{search,resolution}.rs (jamais surfacé
live parce qu'aucun snippet n'avait de score en DB). Surface par
le premier recompute réel sur les 10 repos GitHub seedés, où le
endpoint GET /api/repos/search décodait pour de bon.

Fix : cast `::float8` dans le SELECT. Les scores sont sérialisés
en JSON de toute façon — f64 suffit sur une précision 4,3.

Also: restore default-run = "usestakly-backend" dans Cargo.toml.
Depuis l'ajout du bin seed_github (R1), `cargo run` échouait sur
l'ambiguïté alors que CLAUDE.md documente toujours cette commande
comme point d'entrée standard.

Vérification live : POST /api/admin/scoring/recompute retourne
{externalsProcessed: 10}. GET /api/repos/search?filter=explore
renvoie les 10 repos triés par overall, avec scoring cohérent
(react/next.js fresh=0.99 overall=0.65, moment/request fresh=0.09
overall=0.24). filter=auto/strict à 0 items (reliability=0.5
neutral_default sans signaux build).

+++ backend/src/services/repos.rs
@@ -47,15 +47,15 @@ pub async fn search_github_repos(
           e.forks_count          AS forks_count,
           e.open_issues_count    AS open_issues_count,
           e.archived             AS archived,
-          e.last_commit_at       AS last_commit_at,
-          ascore.formula_version AS quality_formula_version,
-          ascore.freshness       AS quality_freshness,
-          ascore.adoption        AS quality_adoption,
-          ascore.reliability     AS quality_reliability,
-          ascore.abandonment     AS quality_abandonment,
-          ascore.overall         AS quality_overall,
-          ascore.flags           AS quality_flags,
-          ascore.computed_at     AS quality_computed_at
+          e.last_commit_at            AS last_commit_at,
+          ascore.formula_version      AS quality_formula_version,
+          ascore.freshness::float8    AS quality_freshness,
+          ascore.adoption::float8     AS quality_adoption,
+          ascore.reliability::float8  AS quality_reliability,
+          ascore.abandonment::float8  AS quality_abandonment,
+          ascore.overall::float8      AS quality_overall,
+          ascore.flags                AS quality_flags,
+          ascore.computed_at          AS quality_computed_at
         FROM external_artifacts e
         LEFT JOIN artifact_scores ascore
           ON ascore.external_artifact_id = e.id
@@ -135,17 +135,17 @@ pub async fn get_repo_profile(db: &PgPool, artifact_id: Uuid) -> Result<RepoProf
           e.open_issues_count    AS open_issues_count,
           e.archived             AS archived,
           e.last_commit_at       AS last_commit_at,
-          e.subscribers_count    AS subscribers_count,
-          e.default_branch       AS default_branch,
-          e.priors_fetched_at    AS priors_fetched_at,
-          ascore.formula_version AS quality_formula_version,
-          ascore.freshness       AS quality_freshness,
-          ascore.adoption        AS quality_adoption,
-          ascore.reliability     AS quality_reliability,
-          ascore.abandonment     AS quality_abandonment,
-          ascore.overall         AS quality_overall,
-          ascore.flags           AS quality_flags,
-          ascore.computed_at     AS quality_computed_at
+          e.subscribers_count         AS subscribers_count,
+          e.default_branch            AS default_branch,
+          e.priors_fetched_at         AS priors_fetched_at,
+          ascore.formula_version      AS quality_formula_version,
+          ascore.freshness::float8    AS quality_freshness,
+          ascore.adoption::float8     AS quality_adoption,
+          ascore.reliability::float8  AS quality_reliability,
+          ascore.abandonment::float8  AS quality_abandonment,
+          ascore.overall::float8      AS quality_overall,
+          ascore.flags                AS quality_flags,
+          ascore.computed_at          AS quality_computed_at
         FROM external_artifacts e
         LEFT JOIN artifact_scores ascore
           ON ascore.external_artifact_id = e.id
